+++ src/main/java/uk/ac/standrews/cs/utilities/m_tree/experiments/euclidean/Point.java
@@ -18,10 +18,10 @@ package uk.ac.standrews.cs.utilities.m_tree.experiments.euclidean;
 
 public class Point implements Comparable<Point> {
 
-    public final float x;
-    public final float y;
+    public final double x;
+    public final double y;
 
-    public Point(final float x, final float y) {
+    public Point(final double x, final double y) {
         this.x = x;
         this.y = y;
     }
